moved JsonFileStats.java into ThreadStatsCounter.java as inner class. Added documentations to both classes
